mavconn-imagestreamer: lower default quality to '60'
'100' leads to display problems on the qgroundcontrol side and an excessive amount of packets/frame being used (>180 on my setup). Lowering it to '60' fixes the display problems and lowers the packet count to a more reasonable 116 packets/frame. '0' is still the clear winner when it comes to packet count: 21 packets/frame.

And it significantly lowers the cpu usage from 90% to 60% for mavconn-imagestreamer on an 1.2GHz dual-core cortex-A9 system.

Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>